refactor: feed configuration file structure
